* remove all external resource fetches in favor of local * do tail parsing highlighting server side
