+++ islands/tail.tsx
@@ -5,8 +5,6 @@ import IconPlayerPlayFilled from "tabler-icons/tsx/player-play-filled.tsx";
 import IconWindowMinimize from "tabler-icons/tsx/window-minimize.tsx";
 import IconWindowMaximize from "tabler-icons/tsx/window-maximize.tsx";
 import IconX from "tabler-icons/tsx/x.tsx";
-import { Head } from "$fresh/runtime.ts";
-import hljs from "https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/es/highlight.min.js";
 
 import { useEffect, useRef, useState } from "preact/hooks";
 import { signal } from "@preact/signals";
@@ -25,16 +23,6 @@ export const tailSamplingSignal = signal<boolean>(false);
 export const tailSamplingRateSignal = signal<number>(1);
 
 export type TailData = { timestamp: Date; data: string };
-
-export const parseData = (data: string) => {
-  try {
-    const parsed = JSON.parse(data);
-    return JSON.stringify(parsed, null, 2);
-  } catch (e) {
-    console.debug("Error parsing tail data, returning raw data instead");
-  }
-  return data;
-};
 
 export const TailRow = (
   { row }: { row: TailData; index: number },
@@ -65,7 +53,7 @@ export const TailRow = (
         <pre>
           <code>
             <div dangerouslySetInnerHTML={{
-                __html: hljs.highlightAuto(parseData(row.data)).value,
+                __html: row.data,
               }}
             >
             </div>
@@ -88,73 +76,65 @@ export const Tail = ({ audience }: { audience: Audience }) => {
   }, [tailSamplingSignal.value, tailSamplingRateSignal.value]);
 
   return (
-    <>
-      <Head>
-        <link
-          rel="stylesheet"
-          href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/styles/dark.min.css"
-        />
-      </Head>
+    <div
+      class={`relative flex flex-col h-screen w-[calc(100vw-${OP_MODAL_WIDTH})]`}
+    >
+      <div class="h-46 w-full bg-streamdalPurple p-4 text-white font-semibold text-sm">
+        <span class="opacity-50">Home</span> / Tail
+      </div>
       <div
-        class={`relative flex flex-col h-screen w-[calc(100vw-${OP_MODAL_WIDTH})]`}
+        class={`h-full flex flex-col bg-white p-4 ${
+          fullScreen
+            ? "absolute top-0 bottom-0 right-0 left-0 z-[51] w-screen h-screen"
+            : ""
+        }`}
       >
-        <div class="h-46 w-full bg-streamdalPurple p-4 text-white font-semibold text-sm">
-          <span class="opacity-50">Home</span> / Tail
+        <div
+          class={`flex flew-row justify-between item-center mt-6 my-4 mx-auto text-3xl font-medium w-[${
+            fullScreen ? "100" : "90"
+          }%]`}
+        >
+          <div>
+            Tail{" "}
+            <span class="text-streamdalPurple">{audience.operationName}</span>
+          </div>
+          <div class="flex flex-row justify-end items-center">
+            <div
+              class="flex justify-center items-center w-[36px] h-[36px] rounded-[50%] bg-streamdalPurple cursor-pointer"
+              onClick={() => tailPausedSignal.value = !tailPausedSignal.value}
+            >
+              {tailPausedSignal.value
+                ? <IconPlayerPlayFilled class="w-6 h-6 text-white" />
+                : <IconPlayerPauseFilled class="w-6 h-6 text-white" />}
+            </div>
+            <div
+              className="ml-2 flex justify-center items-center w-[36px] h-[36px] rounded-[50%] bg-streamdalPurple cursor-pointer"
+              onClick={() => setFullScreen(!fullScreen)}
+            >
+              {fullScreen
+                ? <IconWindowMinimize class="w-6 h-6 text-white" />
+                : <IconWindowMaximize class="w-6 h-6 text-white" />}
+            </div>
+            <div
+              className="ml-2 flex justify-center items-center w-[36px] h-[36px] rounded-[50%] bg-streamdalPurple cursor-pointer"
+              onClick={() => tailEnabledSignal.value = false}
+            >
+              <IconX class="w-6 h-6 text-white" />
+            </div>
+          </div>
         </div>
         <div
-          class={`h-full flex flex-col bg-white p-4 ${
-            fullScreen
-              ? "absolute top-0 bottom-0 right-0 left-0 z-[51] w-screen h-screen"
-              : ""
-          }`}
+          class={`flex flex-col mx-auto w-[${
+            fullScreen ? "100" : "90"
+          }%] h-[calc(100vh-${
+            fullScreen ? "200" : "260"
+          }px)] overflow-y-scroll rounded-md bg-black text-white`}
         >
-          <div
-            class={`flex flew-row justify-between item-center mt-6 my-4 mx-auto text-3xl font-medium w-[${
-              fullScreen ? "100" : "90"
-            }%]`}
-          >
-            <div>
-              Tail{" "}
-              <span class="text-streamdalPurple">{audience.operationName}</span>
-            </div>
-            <div class="flex flex-row justify-end items-center">
-              <div
-                class="flex justify-center items-center w-[36px] h-[36px] rounded-[50%] bg-streamdalPurple cursor-pointer"
-                onClick={() => tailPausedSignal.value = !tailPausedSignal.value}
-              >
-                {tailPausedSignal.value
-                  ? <IconPlayerPlayFilled class="w-6 h-6 text-white" />
-                  : <IconPlayerPauseFilled class="w-6 h-6 text-white" />}
-              </div>
-              <div
-                className="ml-2 flex justify-center items-center w-[36px] h-[36px] rounded-[50%] bg-streamdalPurple cursor-pointer"
-                onClick={() => setFullScreen(!fullScreen)}
-              >
-                {fullScreen
-                  ? <IconWindowMinimize class="w-6 h-6 text-white" />
-                  : <IconWindowMaximize class="w-6 h-6 text-white" />}
-              </div>
-              <div
-                className="ml-2 flex justify-center items-center w-[36px] h-[36px] rounded-[50%] bg-streamdalPurple cursor-pointer"
-                onClick={() => tailEnabledSignal.value = false}
-              >
-                <IconX class="w-6 h-6 text-white" />
-              </div>
-            </div>
-          </div>
-          <div
-            class={`flex flex-col mx-auto w-[${
-              fullScreen ? "100" : "90"
-            }%] h-[calc(100vh-${
-              fullScreen ? "200" : "260"
-            }px)] overflow-y-scroll rounded-md bg-black text-white`}
-          >
-            {tailSignal.value[audienceKey(audience)]?.map((
-              tail: TailData,
-            ) => <TailRow row={tail} />)}
-          </div>
+          {tailSignal.value[audienceKey(audience)]?.map((
+            tail: TailData,
+          ) => <TailRow row={tail} />)}
         </div>
       </div>
-    </>
+    </div>
   );
 };
